refactor(frontend): extract ProjectMenu and GlobalMenu from RootLayout
Move the self-contained ProjectMenu and GlobalMenu topbar menu components
(with their props interfaces) out of the large RootLayout file into their
own modules, and move the shared ACTION_MENU_ITEM_ICON_SX /
ACTION_MENU_ICON_PROPS constants into a shared topbarMenuStyles module.
RootLayout drops the icon imports that were only used by the menus. No
behavior change.
