+++ frontend/src/navigation/topbarMenuStyles.ts
@@ -1,0 +1,5 @@
+// Shared styling for the topbar action menus (ProjectMenu, GlobalMenu):
+// the leading icon column and the icon size used inside their menu items.
+
+export const ACTION_MENU_ITEM_ICON_SX = { minWidth: 32, color: 'text.secondary' } as const;
+export const ACTION_MENU_ICON_PROPS = { fontSize: 'small' } as const;
